requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.2.2
python-dotenv==1.0.0
supabase==2.8.0
pydantic==2.5.2
python-multipart==0.0.6
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0
